# S5.01_API_REST/fitbit_api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitbit API Documentation",
        "description": "API para gestión de disciplinas, comunidades y usuarios",
        "contact": {
            "email": "admin@fitbit.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/stats/communities": {
            "get": {
                "tags": [
                    "Community Stats"
                ],
                "summary": "Estadísticas generales de comunidades",
                "description": "Obtiene estadísticas generales de las comunidades",
                "operationId": "getCommunityStats",
                "responses": {
                    "200": {
                        "description": "Estadísticas obtenidas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "total_communities": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total_users": {
                                                    "type": "integer",
                                                    "example": 120
                                                },
                                                "most_popular_community": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Boxing Community"
                                                        },
                                                        "users_count": {
                                                            "type": "integer",
                                                            "example": 25
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "communities_per_discipline": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "discipline": {
                                                                "type": "string",
                                                                "example": "Boxing"
                                                            },
                                                            "communities_count": {
                                                                "type": "integer",
                                                                "example": 5
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "403": {
                        "description": "Sin permisos para ver estadísticas"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/stats/communities/ranking": {
            "get": {
                "tags": [
                    "Community Stats"
                ],
                "summary": "Ranking de comunidades por miembros",
                "description": "Obtiene el top 10 de comunidades con más miembros",
                "operationId": "getCommunityRanking",
                "responses": {
                    "200": {
                        "description": "Ranking obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Boxing Community"
                                                    },
                                                    "users_count": {
                                                        "type": "integer",
                                                        "example": 25
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/stats/communities/percentage": {
            "get": {
                "tags": [
                    "Community Stats"
                ],
                "summary": "Porcentaje de usuarios por comunidad",
                "description": "Obtiene el porcentaje de distribución de usuarios en comunidades",
                "operationId": "getCommunityPercentage",
                "responses": {
                    "200": {
                        "description": "Porcentajes obtenidos exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "community_name": {
                                                        "type": "string",
                                                        "example": "Boxing Community"
                                                    },
                                                    "users_count": {
                                                        "type": "integer",
                                                        "example": 25
                                                    },
                                                    "percentage": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 20.83
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/stats/communities/summary": {
            "get": {
                "tags": [
                    "Community Stats"
                ],
                "summary": "Resumen completo de comunidades",
                "description": "Obtiene un resumen detallado de estadísticas de comunidades",
                "operationId": "getCommunitySummary",
                "responses": {
                    "200": {
                        "description": "Resumen obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "total_communities": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "communities_with_moderators": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "communities_without_moderators": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "average_members_per_community": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 8.5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/stats/communities/by-discipline": {
            "get": {
                "tags": [
                    "Community Stats"
                ],
                "summary": "Comunidades agrupadas por disciplina",
                "description": "Obtiene el número de comunidades por cada disciplina",
                "operationId": "getCommunitiesByDiscipline",
                "responses": {
                    "200": {
                        "description": "Estadísticas por disciplina obtenidas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "discipline_name": {
                                                        "type": "string",
                                                        "example": "Boxing"
                                                    },
                                                    "communities_count": {
                                                        "type": "integer",
                                                        "example": 5
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "type": "http",
            "description": "HTTP authentication",
            "name": "Authorization",
            "in": "header",
            "scheme": "bearer"
        }
    },
...
    "security": [
        {
            "bearer_token": []
        }
    ]
}